Scope poll authoring to /author
