+++ app/views/author/polls/_empty_state.html.erb
@@ -4,7 +4,7 @@
   </p>
 
   <%= link_to "Add poll",
-        new_poll_path,
+        new_author_poll_path,
         class: "button primary",
         data: {turbo_frame: dom_id(Poll.new)} %>
 </div>
