+++ app/views/author/polls/_form.html.erb
@@ -7,6 +7,6 @@
 
   <%= f.label :title %>
   <%= f.text_field :title, autofocus: true %>
-  <%= link_to "Cancel", polls_path, class: "button transparent" %>
+  <%= link_to "Cancel", author_polls_path, class: "button transparent" %>
   <%= f.submit class: "button secondary" %>
 <% end %>
