+++ app/views/author/polls/_poll.html.erb
@@ -1,15 +1,15 @@
 <%= turbo_frame_tag poll do %>
   <div class="poll">
     <%= link_to poll.title,
-          poll_path(poll),
+          author_poll_path(poll),
           data: {turbo_frame: "_top"} %>
     <div class="poll__actions">
       <%= button_to "Delete",
-            poll_path(poll),
+            author_poll_path(poll),
             method: :delete,
             class: "button secondary" %>
       <%= link_to "Edit",
-            edit_poll_path(poll),
+            edit_author_poll_path(poll),
             class: "button secondary" %>
     </div>
   </div>
