+++ app/views/author/polls/_total.html.erb
