+++ app/views/author/polls/answers/create.turbo_stream.erb
@@ -1,11 +1,11 @@
 <%= turbo_stream.update nested_dom_id(@question, Polls::Answer.new), "" %>
 
 <%= turbo_stream.append nested_dom_id(@question, "answers") do %>
-  <%= render @answer, poll: @poll, question: @question %>
+  <%= render "author/polls/answers/answer", answer: @answer, poll: @poll, question: @question %>
 <% end %>
 
 <%= turbo_stream.update dom_id(@poll, :total) do %>
-  <%= render "polls/total", poll: @poll %>
+  <%= render "author/polls/total", poll: @poll %>
 <% end %>
 
 <%= turbo_stream.prepend "flash", partial: "application/flash" %>
