+++ app/views/author/polls/answers/destroy.turbo_stream.erb
