+++ app/views/author/polls/answers/new.html.erb
@@ -9,5 +9,5 @@
           answer: @answer %>
   <% end %>
 
-  <%= link_to sanitize("&larr; Back to poll"), poll_path(@poll) %>
+  <%= link_to sanitize("&larr; Back to poll"), author_poll_path(@poll) %>
 </div>
