+++ app/views/author/polls/destroy.turbo_stream.erb
