+++ app/views/author/polls/index.html.erb
@@ -5,7 +5,7 @@
 <div class="section-content container py-gap mb-3xl">
   <div class="header">
     <%= link_to "New poll",
-          new_poll_path,
+          new_author_poll_path,
           class: "button primary",
           data: {turbo_frame: dom_id(Poll.new)} %>
   </div>
