+++ app/views/author/polls/new.html.erb
@@ -1,8 +1,8 @@
-<%= render Pages::Header.new(title: "Edit Poll") %>
+<%= render Pages::Header.new(title: "New Poll") %>
 <div class="section-content container py-gap mb-3xl">
   <%= turbo_frame_tag @poll do %>
     <%= render "form", poll: @poll %>
   <% end %>
 
-  <%= link_to sanitize("&larr; Back to poll"), poll_path(@poll) %>
+  <%= link_to sanitize("&larr; Back to polls"), author_polls_path %>
 </div>
