+++ app/views/author/polls/questions/_question.html.erb
@@ -8,12 +8,12 @@
 
         <div class="question__actions">
           <%= button_to "Delete",
-                poll_question_path(poll, question),
+                author_poll_question_path(poll, question),
                 method: :delete,
                 form: {data: {turbo_confirm: "Are you sure?"}},
                 class: "button transparent" %>
           <%= link_to "Edit",
-                edit_poll_question_path(poll, question),
+                edit_author_poll_question_path(poll, question),
                 class: "button transparent" %>
         </div>
       </div>
@@ -31,7 +31,7 @@
 
       <div class="question__footer">
         <%= link_to "New answer",
-              new_poll_question_answer_path(poll, question),
+              new_author_poll_question_answer_path(poll, question),
               data: {turbo_frame: nested_dom_id(question, Polls::Answer.new)},
               class: "button primary" %>
       </div>
