+++ app/views/author/polls/questions/destroy.turbo_stream.erb
